# [Go] Zero out the slice at the end of the hot loop