@jinyus 2023-11-20
Results from new VM:

Main: `19.10 ms | 282.11 ms | 2.48 s `

This PR:
```
Go:

        Processing time (w/o IO): 19.31218ms
        total: 0.04s memory: 13128k
        Processing time (w/o IO): 18.959362ms
        total: 0.04s memory: 12996k
        Processing time (w/o IO): 18.944572ms
        total: 0.04s memory: 13128k
        Processing time (w/o IO): 19.166462ms
        total: 0.04s memory: 13052k
        Processing time (w/o IO): 19.235371ms
        total: 0.04s memory: 13004k
        Processing time (w/o IO): 18.955822ms
        total: 0.04s memory: 13512k
        Processing time (w/o IO): 19.24913ms
        total: 0.04s memory: 13168k
        Processing time (w/o IO): 19.026742ms
        total: 0.04s memory: 13012k
        Processing time (w/o IO): 19.054162ms
        total: 0.04s memory: 13008k
        Processing time (w/o IO): 19.18756ms
        total: 0.04s memory: 13008k

Go:

        Processing time (w/o IO): 279.608796ms
        total: 0.39s memory: 44436k
        Processing time (w/o IO): 282.393582ms
        total: 0.40s memory: 44564k
        Processing time (w/o IO): 281.712966ms
        total: 0.39s memory: 44080k

Go:

        Processing time (w/o IO): 2.48916449s
        total: 2.91s memory: 161356k
        Processing time (w/o IO): 2.476178139s
        total: 2.88s memory: 161508k
        Processing time (w/o IO): 2.464193559s
        total: 2.87s memory: 160032k
```